+++ libraries/common/src/test/java/androidx/media3/common/audio/AudioProcessingPipelineTest.java
@@ -47,7 +47,7 @@ public final class AudioProcessingPipelineTest {
         new AudioProcessingPipeline(ImmutableList.of());
 
     audioProcessingPipeline.configure(AUDIO_FORMAT);
-    audioProcessingPipeline.flush();
+    audioProcessingPipeline.flush(StreamMetadata.DEFAULT);
 
     assertThat(audioProcessingPipeline.isOperational()).isFalse();
   }
@@ -63,7 +63,7 @@ public final class AudioProcessingPipelineTest {
             ImmutableList.of(audioProcessorOne, audioProcessorTwo, audioProcessorThree));
     // The internal state of the pipeline does not affect equality.
     pipelineOne.configure(AUDIO_FORMAT);
-    pipelineOne.flush();
+    pipelineOne.flush(StreamMetadata.DEFAULT);
 
     AudioProcessingPipeline pipelineTwo =
         new AudioProcessingPipeline(
@@ -123,7 +123,7 @@ public final class AudioProcessingPipelineTest {
     audioProcessingPipeline.configure(AUDIO_FORMAT);
     // Configuring the pipeline is not enough for it to be operational.
     assertThat(audioProcessingPipeline.isOperational()).isFalse();
-    audioProcessingPipeline.flush();
+    audioProcessingPipeline.flush(StreamMetadata.DEFAULT);
     assertThat(audioProcessingPipeline.isOperational()).isTrue();
   }
 
@@ -133,13 +133,13 @@ public final class AudioProcessingPipelineTest {
     AudioProcessingPipeline audioProcessingPipeline =
         new AudioProcessingPipeline(ImmutableList.of(audioProcessor));
     audioProcessingPipeline.configure(AUDIO_FORMAT);
-    audioProcessingPipeline.flush();
+    audioProcessingPipeline.flush(StreamMetadata.DEFAULT);
     assertThat(audioProcessingPipeline.isOperational()).isTrue();
 
     audioProcessor.setActive(false);
     audioProcessingPipeline.configure(AUDIO_FORMAT);
     assertThat(audioProcessingPipeline.isOperational()).isTrue();
-    audioProcessingPipeline.flush();
+    audioProcessingPipeline.flush(StreamMetadata.DEFAULT);
     assertThat(audioProcessingPipeline.isOperational()).isFalse();
   }
 
@@ -162,7 +162,7 @@ public final class AudioProcessingPipelineTest {
     AudioProcessingPipeline audioProcessingPipeline =
         new AudioProcessingPipeline(ImmutableList.of(fakeSampleRateChangingAudioProcessor));
     AudioFormat outputFormat = audioProcessingPipeline.configure(AUDIO_FORMAT);
-    audioProcessingPipeline.flush();
+    audioProcessingPipeline.flush(StreamMetadata.DEFAULT);
     assertThat(outputFormat).isEqualTo(AUDIO_FORMAT);
     assertThat(audioProcessingPipeline.isOperational()).isFalse();
   }
@@ -173,7 +173,7 @@ public final class AudioProcessingPipelineTest {
     AudioProcessingPipeline audioProcessingPipeline =
         new AudioProcessingPipeline(ImmutableList.of(audioProcessor));
     audioProcessingPipeline.configure(AUDIO_FORMAT);
-    audioProcessingPipeline.flush();
+    audioProcessingPipeline.flush(StreamMetadata.DEFAULT);
 
     ByteBuffer inputBuffer = createOneSecondDefaultSilenceBuffer(AUDIO_FORMAT);
     long inputBytes = inputBuffer.remaining();
@@ -190,7 +190,7 @@ public final class AudioProcessingPipelineTest {
     AudioProcessingPipeline audioProcessingPipeline =
         new AudioProcessingPipeline(ImmutableList.of(audioProcessor));
     audioProcessingPipeline.configure(AUDIO_FORMAT);
-    audioProcessingPipeline.flush();
+    audioProcessingPipeline.flush(StreamMetadata.DEFAULT);
 
     ByteBuffer inputBuffer = createOneSecondDefaultSilenceBuffer(AUDIO_FORMAT);
     audioProcessingPipeline.queueInput(inputBuffer);
@@ -217,7 +217,7 @@ public final class AudioProcessingPipelineTest {
         };
     AudioProcessingPipeline pipeline = new AudioProcessingPipeline(ImmutableList.of(processor));
     pipeline.configure(AUDIO_FORMAT);
-    pipeline.flush(new StreamMetadata(/* positionOffsetUs= */ 0));
+    pipeline.flush(StreamMetadata.DEFAULT);
     assertThat(lastPositionOffsetUs.get()).isEqualTo(0);
   }
 
@@ -235,7 +235,7 @@ public final class AudioProcessingPipelineTest {
         };
     AudioProcessingPipeline pipeline = new AudioProcessingPipeline(ImmutableList.of(processor));
     pipeline.configure(AUDIO_FORMAT);
-    pipeline.flush();
+    pipeline.flush(StreamMetadata.DEFAULT);
     assertThat(lastPositionOffsetUs.get()).isEqualTo(0);
   }
 
@@ -281,7 +281,7 @@ public final class AudioProcessingPipelineTest {
         new AudioProcessingPipeline(
             ImmutableList.of(firstProcessor, secondProcessor, thirdProcessor));
     pipeline.configure(AUDIO_FORMAT);
-    pipeline.flush(new StreamMetadata(/* positionOffsetUs= */ 1_000_000));
+    pipeline.flush(new StreamMetadata.Builder().setPositionOffsetUs(1_000_000).build());
     assertThat(firstProcessorLastPositionOffset.get()).isEqualTo(/* positionOffsetUs */ 1_000_000);
     // The first AudioProcessor doubles the speed of the input stream, so the offset needs to adjust
     // accordingly.
@@ -305,7 +305,7 @@ public final class AudioProcessingPipelineTest {
                     /* maxInputBytesAtOnce= */ 160,
                     /* duplicateBytes= */ false)));
     audioProcessingPipeline.configure(AUDIO_FORMAT);
-    audioProcessingPipeline.flush();
+    audioProcessingPipeline.flush(StreamMetadata.DEFAULT);
 
     ByteBuffer inputBuffer = createOneSecondDefaultSilenceBuffer(AUDIO_FORMAT);
     inputBuffer.put(0, (byte) 24);